refactor: 팀 체크리스트 편집 화면 패키지 구조 변경 및 items 로직 변경
