+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/edit/team/TeamChecklistEditScreen.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.compose.edit.team.main
+package com.bottari.presentation.compose.edit.team
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -8,13 +8,13 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalFocusManager
 import com.bottari.presentation.compose.common.modifier.noRippleClickable
 import com.bottari.presentation.compose.common.theme.BottariTheme
 import com.bottari.presentation.compose.edit.personal.item.component.ItemEditInputBar
-import com.bottari.presentation.compose.edit.team.TeamChecklistItem
 import com.bottari.presentation.model.bottari.BottariItemUiModel
 
 @OptIn(ExperimentalLayoutApi::class)
@@ -46,8 +46,8 @@ fun TeamChecklistEditScreen(
                     .padding(top = BottariTheme.spacing.spaceSmall)
                     .weight(1f),
         ) {
-            items(items.size) { index ->
-                TeamChecklistItem(item = items[index], onDeleteClick = onDeleteItem)
+            items(items) { item ->
+                TeamChecklistItem(item = item, onDeleteClick = onDeleteItem)
             }
         }
         ItemEditInputBar(
